+++ pluralize/is.js
@@ -1,3 +1,1 @@
-module.exports = function(n) {
-  return n % 10 !== 1 || n % 100 === 11 ? 'p' : 's'
-}
+module.exports = require('./').plural_IS